requirements.txt
attrs==23.1.0
Automat==22.10.0
beautifulsoup4==4.12.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
constantly==15.1.0
contourpy==1.1.0
cryptography==41.0.1
cssselect==1.2.0
cycler==0.11.0
filelock==3.12.0
fonttools==4.40.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
jmespath==1.0.1
joblib==1.2.0
kiwisolver==1.4.4
lxml==4.9.2
matplotlib==3.7.1
mysql-connector-python==8.0.33
numpy==1.24.3
packaging==23.1
pandas==2.0.2
parsel==1.8.1
Pillow==9.5.0
Protego==0.2.1
protobuf==3.20.3
proxyscrape==0.3.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
pyOpenSSL==23.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
queuelib==1.6.2
requests==2.31.0
requests-file==1.5.1
scikit-learn==1.2.2
scipy==1.10.1
Scrapy==2.9.0
scrapy-user-agents==0.1.1
service-identity==21.1.0
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.16
threadpoolctl==3.1.0
tldextract==3.4.4
Twisted==22.10.0
typing==3.7.4.3
typing_extensions==4.6.3
tzdata==2023.3
ua-parser==0.16.1
urllib3==2.0.3
user-agents==2.2.0
w3lib==2.1.1
zope.interface==6.0
